Update MotorHat project
