+++ src/devices/ServoMotor/ServoMotor.cs
@@ -99,7 +99,7 @@ namespace Iot.Device.ServoMotor
         public void Dispose()
         {
             _pwmChannel?.Dispose();
-            _pwmChannel = null;
+            _pwmChannel = null!;
         }
     }
 }
